[Edit] 코드 정리 (#19)
