+++ 29cm-iOS/29cm-iOS/Source/Common/Extensions/UIView+.swift
@@ -32,4 +32,23 @@ extension UIView {
         
         self.layer.mask = shapeLayer
     }
+    
+    // @IBInspectable
+    @IBInspectable var borderWidth: CGFloat {
+        get {
+            return layer.borderWidth
+        }
+        set {
+            layer.borderWidth = newValue
+        }
+    }
+    
+    @IBInspectable var borderColor: UIColor? {
+        get {
+            return UIColor(cgColor: layer.borderColor!)
+        }
+        set {
+            layer.borderColor = newValue?.cgColor
+        }
+    }
 }
