+++ 29cm-iOS/29cm-iOS/Source/Common/Protocols/NavigationBarViewDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  NavigationBarViewDelegate.swift
+//  29cm-iOS
+//
+//  Created by 정은희 on 2022/05/30.
+//
+
+import Foundation
+
+protocol NavigationBarViewDelegate {
+    func dismissButtonClicked()
+}
