+++ 29cm-iOS/29cm-iOS/Source/Presentation/Create/Views/Cells/Hee/CheckListCollectionViewCell.swift
@@ -33,8 +33,6 @@ final class CheckListCollectionViewCell: UICollectionViewCell {
     // MARK: - View Life Cycle
     override func awakeFromNib() {
         super.awakeFromNib()
-        
-        setStyle()
     }
     
     // MARK: - @IBAction Properties
@@ -59,15 +57,6 @@ final class CheckListCollectionViewCell: UICollectionViewCell {
 }
 
 // MARK: - Extensions
-extension CheckListCollectionViewCell {
-    // TODO: - setStyle() → @IBInspectable 이용하는 것으로 변경
-    private func setStyle() {
-        [numberTextField, numberButton].forEach {
-            $0?.makeRoundedWithBorder(radius: 0, borderColor: Const.Color.cmBoxGrey!.cgColor)
-        }
-    }
-}
-
 extension CheckListCollectionViewCell: UITextFieldDelegate {
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
         textField.resignFirstResponder()
