+++ 29cm-iOS/29cm-iOS/Source/Presentation/Create/Views/CompletionPopUpView.storyboard
@@ -59,6 +59,14 @@
                                                 <state key="normal" title="확인">
                                                     <color key="titleColor" name="grey01_29cm"/>
                                                 </state>
+                                                <userDefinedRuntimeAttributes>
+                                                    <userDefinedRuntimeAttribute type="number" keyPath="borderWidth">
+                                                        <real key="value" value="0.0"/>
+                                                    </userDefinedRuntimeAttribute>
+                                                    <userDefinedRuntimeAttribute type="color" keyPath="borderColor">
+                                                        <color key="value" name="line_grey_29cm"/>
+                                                    </userDefinedRuntimeAttribute>
+                                                </userDefinedRuntimeAttributes>
                                                 <connections>
                                                     <action selector="completionButtonClicked:" destination="Y6W-OH-hqX" eventType="touchUpInside" id="Erf-lW-jeg"/>
                                                 </connections>
@@ -75,6 +83,14 @@
                                     <constraint firstItem="aTQ-kA-7dJ" firstAttribute="top" secondItem="d0d-Rl-ubA" secondAttribute="top" constant="27" id="g0b-H8-w9m"/>
                                     <constraint firstItem="tkD-OR-xyp" firstAttribute="top" secondItem="d0d-Rl-ubA" secondAttribute="top" constant="45" id="vC0-UC-aOa"/>
                                 </constraints>
+                                <userDefinedRuntimeAttributes>
+                                    <userDefinedRuntimeAttribute type="number" keyPath="borderWidth">
+                                        <real key="value" value="0.0"/>
+                                    </userDefinedRuntimeAttribute>
+                                    <userDefinedRuntimeAttribute type="color" keyPath="borderColor">
+                                        <color key="value" name="black01_29cm"/>
+                                    </userDefinedRuntimeAttribute>
+                                </userDefinedRuntimeAttributes>
                             </view>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="vDu-zF-Fre"/>
@@ -105,6 +121,9 @@
         </namedColor>
         <namedColor name="grey01_29cm">
             <color red="0.37254901960784315" green="0.37254901960784315" blue="0.37254901960784315" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
+        <namedColor name="line_grey_29cm">
+            <color red="0.89411764705882357" green="0.89411764705882357" blue="0.89411764705882357" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </namedColor>
         <systemColor name="systemBackgroundColor">
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
